requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
async-timeout==4.0.3
attrs==25.4.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.3
dataclasses-json==0.6.7
decorator==5.2.1
exceptiongroup==1.3.0
faiss-cpu==1.12.0
filelock==3.19.1
filetype==1.2.0
frozenlist==1.8.0
fsspec==2025.9.0
google-ai-generativelanguage==0.6.15
google-api-core==2.26.0
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
imageio==2.37.0
imageio-ffmpeg==0.6.0
instaloader==4.14.2
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.31
langchain-core==0.3.79
langchain-google-genai==2.0.10
langchain-text-splitters==0.3.11
langsmith==0.4.34
llvmlite==0.43.0
MarkupSafe==3.0.3
marshmallow==3.26.1
more-itertools==10.8.0
moviepy==2.2.1
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
openai-whisper==20250625
opencv-python==4.12.0.88
orjson==3.11.3
packaging==25.0
pillow==11.3.0
proglog==0.1.12
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.0
pydantic-settings==2.11.0
pydantic_core==2.41.1
pyparsing==3.2.5
python-dotenv==1.1.1
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
sniffio==1.3.1
SQLAlchemy==2.0.44
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.12.0
torch==2.8.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
yarl==1.22.0
zstandard==0.25.0
